Extract RealmFind & RealmFindAll commands. Subscribe to all save commands.
