+++ app/src/main/java/io/ipoli/android/quest/commands/StartQuestCommand.java
@@ -52,7 +52,7 @@ public class StartQuestCommand {
             for (Quest cq : quests) {
                 if (!cq.getId().equals(q.getId()) && Quest.isStarted(cq)) {
                     cq.setActualStart(null);
-                    questPersistenceService.save(cq);
+                    questPersistenceService.save(cq).subscribe();
                 }
             }
         });
