+++ app/src/main/java/io/ipoli/android/quest/commands/StopQuestCommand.java
@@ -2,8 +2,8 @@ package io.ipoli.android.quest.commands;
 
 import android.content.Context;
 
-import io.ipoli.android.quest.data.Quest;
 import io.ipoli.android.quest.QuestNotificationScheduler;
+import io.ipoli.android.quest.data.Quest;
 import io.ipoli.android.quest.persistence.QuestPersistenceService;
 
 /**
@@ -24,7 +24,8 @@ public class StopQuestCommand {
 
     public void execute() {
         quest.setActualStart(null);
-        questPersistenceService.save(quest);
-        QuestNotificationScheduler.stopAll(quest.getId(), context);
+        questPersistenceService.save(quest).subscribe(q -> {
+            QuestNotificationScheduler.stopAll(q.getId(), context);
+        });
     }
 }
